+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 Next
 ====
 
+- Allow for custom expiration time in the generic JSON/XML adapters (#429)
+
 Version 1.2.15 - 2024-02-13
 ===========================